planning

+++ opticalFlow-core/source/opticalFlow/opticalFlow.h
@@ -4,5 +4,5 @@
 namespace opticalFlow
 {
 	//TODO add Mask
-	float* calcOpticalFlow(uint8_t* const pixels, uint32_t width, uint32_t heigth);
+	float* calcOpticalFlow(float* const pixels, uint32_t width, uint32_t heigth);
 }
